Moves upsells to alert-center component
