+++ packages/js/src/dashboard/app.js
@@ -7,8 +7,6 @@ import { SidebarNavigation, useSvgAria } from "@yoast/ui-library";
 import PropTypes from "prop-types";
 import { Link, Route, Routes, useLocation } from "react-router-dom";
 import { MenuItemLink, YoastLogo } from "../shared-admin/components";
-import { PremiumUpsellList } from "../shared-admin/components/premium-upsell-list";
-import SidebarRecommendations from "./components/sidebar-recommendations";
 import { useSelectDashboard } from "./hooks";
 import { FirstTimeConfiguration, AlertCenter } from "./routes";
 
@@ -52,10 +50,6 @@ Menu.propTypes = {
  */
 const App = () => {
 	const { pathname } = useLocation();
-	const isPremium = useSelectDashboard( "selectPreference", [], "isPremium" );
-	const premiumLink = useSelectDashboard( "selectLink", [], "https://yoa.st/17h" );
-	const premiumUpsellConfig = useSelectDashboard( "selectUpsellSettingsAsProps" );
-	const promotions = useSelectDashboard( "selectPreference", [], "promotions", [] );
 	return (
 		<SidebarNavigation activePath={ pathname }>
 			<SidebarNavigation.Mobile
@@ -92,13 +86,7 @@ const App = () => {
 								</Routes>
 							</Transition>
 						</main>
-						{ ! isPremium && <PremiumUpsellList
-							premiumLink={ premiumLink }
-							premiumUpsellConfig={ premiumUpsellConfig }
-							promotions={ promotions }
-						/> }
 					</div>
-					<SidebarRecommendations />
 				</div>
 			</div>
 		</SidebarNavigation>
